+++ akshare/__init__.py
@@ -3123,9 +3123,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.17.14 fix: fix forex_spot_em interface
 1.17.15 fix: fix stock_comment_em interface
 1.17.16 fix: fix bond_zh_cov interface
+1.17.17 fix: fix stock_hk_spot interface
 """
 
-__version__ = "1.17.16"
+__version__ = "1.17.17"
 __author__ = "AKFamily"
 
 import sys