+++ docs/changelog.md
@@ -81,6 +81,11 @@
 
 ## 更新说明详情
 
+1.17.17 fix: fix stock_hk_spot interface
+
+	1. 修复 stock_hk_spot 接口
+	2. 修复 stock_hk_spot 接口
+
 1.17.16 fix: fix bond_zh_cov interface
 
 	1. 修复 bond_zh_cov 接口
@@ -5651,6 +5656,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.17.17 fix: fix stock_hk_spot interface
 
 1.17.16 fix: fix bond_zh_cov interface